+++ cxzmvc/src/main/java/com/cxz/mapper/TestDB/OrderMapper.java
@@ -1,10 +1,9 @@
-package com.cxz.mapper;
+package com.cxz.mapper.TestDB;
 
 
 import com.baomidou.mybatisplus.core.mapper.BaseMapper;
 import com.cxz.model.Order;
 import org.apache.ibatis.annotations.*;
-import org.springframework.beans.factory.annotation.Qualifier;
 
 import java.util.List;
 
@@ -18,11 +17,11 @@ import java.util.List;
  * @since 2021-01-17
  */
 @Mapper
-@Qualifier("sqlSessionFactory")
+//@Qualifier("sqlSessionFactory")
 public interface OrderMapper extends BaseMapper<Order>{
 
 
-    @SelectProvider(type =OrderQueryManager.class,method = "selectAll")
+    @SelectProvider(type = OrderQueryManager.class,method = "selectAll")
 //    @ResultMap("BaseResultMap")
     @Results({@Result(column = "Id",property = "Id"),
             @Result(column = "CreateTime",property = "createtime"),
