+++ cxzmvc/src/main/resources/mapper/TestDB/TUserMapper.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.cxz.mapper.TUserMapper">
+<mapper namespace="com.cxz.mapper.TestDB.TUserMapper">
   <resultMap id="BaseResultMap" type="com.cxz.model.TUser">
     <id column="Id" jdbcType="BIGINT" property="id" />
     <result column="Name" jdbcType="VARCHAR" property="name" />
