+++ cxzmvc/src/main/resources/mapper/TestDB2/TOrderRefundMapper.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.cxz.mapper.TOrderRefundMapper">
+<mapper namespace="com.cxz.mapper.TestDB2.TOrderRefundMapper">
 
                 <!-- 通用查询映射结果 -->
         <resultMap id="BaseResultMap" type="com.cxz.model.TOrderRefund">
